Wrap poll_events in a single transaction to avoid feedback cycles.
## Problem

The problem is that an window-event changes the value of an entity
attribute and the reactive observers on the entity attributes initiates
state changes on the window.

Again a window-event happens and sets the value of an entity attribute.
In the second turn the observer does (or should do) nothing, because the
value of the entity attribute hasn't changed (true -> false -> false),
because it has the same value as before (false -> false).

But there's an exception: if another window-event has been processed and
also set the value of the entity attribute, the value flips between two
states and more and more events are generated (e.g. event1: true ->
false, event2: false -> true, ...).

## Solution

By wrapping all events in a single reactive transaction all state
changes on the entity attributes are collected before the next reactive
turn. As multiple window-events are processed the same entity attribute
could be changed multiple times during the transaction. And it could be
that the value of an entity attribute flipped (e.g. from false to true
and back to false).

Because the changes on the value of an entity attribute isn't propagated
immediately anymore the flipped value will not cause an observer to
react. For the observer there's no any change because the value in the
turn before (e.g. false) and the value in the next turn (e.g. false) are
the same. The immediate values doesn't care. The observer only reacts if
the value is still altered after all window-events has been processed.